Remove unused plugin-specific PDB config code
There was a time when the ASE ArkServerApi provided offsets from the PDB
for only specific structures. ArkServerApi allowed plugins to provide a
config file specifying additional structures they needed offsets for.
ArkServerApi was modified to cache offsets for all structures, but some
code related to plugin-specific PDB configs remained in the codebase
unused. This change removes that unused code.